Move testing classes into testing module
