+++ scoping/src/androidTest/java/com/dropbox/kaiken/scoping/AuthAwareActivitiesTest.kt
@@ -5,7 +5,7 @@ import androidx.test.core.app.ActivityScenario
 import com.google.common.truth.Truth.assertThat
 import org.junit.After
 import org.junit.Test
-import test.com.dropbox.kaiken.scoping.TestAuthAwareScopedActivity
+import test.com.dropbox.kaiken.testing.TestAuthAwareScopedActivity
 
 class AuthAwareActivitiesTest {
     private lateinit var scenario: ActivityScenario<out TestAuthAwareScopedActivity>
