--- scoping/src/debug/java/test/com/dropbox/kaiken/scoping/TestAuthRequireBroadcastReceived.kt
@@ -1,1 +1,0 @@
-package test.com.dropbox.kaiken.scoping
