+++ testing/src/debug/java/test/com/dropbox/kaiken/testing/MyDependencies.kt
@@ -1,4 +1,4 @@
-package test.com.dropbox.kaiken.scoping
+package test.com.dropbox.kaiken.testing
 
 interface MyDependencies {
     fun helloWorldGreeter(): HelloWorldGreeter
