+++ testing/src/debug/java/test/com/dropbox/kaiken/testing/TestActivity.kt
@@ -1,4 +1,4 @@
-package test.com.dropbox.kaiken.scoping
+package test.com.dropbox.kaiken.testing
 
 import android.content.Context
 import android.content.ContextWrapper
