+++ testing/src/debug/java/test/com/dropbox/kaiken/testing/TestAuthAwareBroadcastReceiver.kt
@@ -1,4 +1,4 @@
-package test.com.dropbox.kaiken.scoping
+package test.com.dropbox.kaiken.testing
 
 import android.content.BroadcastReceiver
 import android.content.Context
