+++ testing/src/debug/java/test/com/dropbox/kaiken/testing/TestAuthRequireBroadcastReceived.kt
@@ -1,0 +1,1 @@
+package test.com.dropbox.kaiken.testing
