+++ testing/src/debug/java/test/com/dropbox/kaiken/testing/TestSimpleActivity.kt
@@ -1,4 +1,4 @@
-package test.com.dropbox.kaiken.scoping
+package test.com.dropbox.kaiken.testing
 
 /**
  * Simple test activity that does not inherit from any auth aware activities.
